EmPoWering: gates and figures for the strategy study, and an overflow the figures caught
Thirty-nine new gates, bringing the suite to 139. The strategy ones check what the study
actually rests on: that the mining groups share one hashrate vector and the endowed groups
one stake vector, that disabling a group does not perturb the survivors' draws, that every
group runs alone, that no miner is slower than the Raspberry Pi the distribution is floored
on, that proof-of-work credit reconciles with what the pool paid net of claim fees, and that
every produced block has exactly one leader. The emission gates pin the control function at
both limits -- 62500/657 LGO at maximum, exactly the block's own burn at minimum -- and
confirm the thirty percent target and the sixty-forty split with its sub-two-unit residue.

Three figures, on the validated five-slot categorical palette assigned in fixed order by
strategy so a run with fewer groups cannot repaint the survivors. Composition of a median
node's income by source; accumulated reward per node as a rank curve, which shows dispersion
where a bar would not; and the proof-of-work reward per block and per epoch as two panels
rather than two axes.

Rendering them and looking found two things a validator cannot. The first pass had the
subtitle sitting on the title's descenders, segment labels colliding with both the totals and
the axis labels, and unformatted tick numbers -- all fixed, and the reason the subtitle now
sits above the title rather than between it and the plot is recorded where someone would
otherwise undo it.

The second was a real bug and not a layout one. The per-block reward series accumulated in
int32, and a block's reward is the claim count times a reward of order ten to the ninth base
units, which overflows at ten claims and wraps NEGATIVE. The histogram rendered symmetric
about zero, which is what gave it away. Corrected to int64 and gated three ways: no negative
block reward, the series reconciles with claims times reward, and it carries enough distinct
values to be a distribution at all.

With it fixed the distributions are what they should be: per block a Poisson shape with a
median of 8.4 LGO against a target of ten claims at the opening reward, and per epoch a
spread from 140,000 down through 250,000 LGO carrying the reward's decay.

+++ tools/simulators/empowering/src/empowering_sim/strategies.py
@@ -213,7 +213,10 @@ def run(cfg: Config, scfg: StrategyConfig) -> tuple[Population, list[EpochRecord
 
         # ---- proof of work: the claim flow, block by block, with the retarget in the loop
         hashrate = float(pop.hashrate[pop.mask(*MINING)].sum())
-        claims_found = np.zeros(cfg.blocks_per_epoch, dtype=np.int32)
+        # int64, not int32: a block's reward is the claim count times a reward of order
+        # 10^9 base units, which overflows a 32-bit accumulator at ten claims and wraps
+        # NEGATIVE. The histograms rendered symmetric about zero, which is what caught it.
+        claims_found = np.zeros(cfg.blocks_per_epoch, dtype=np.int64)
         target = state.difficulty_target
         pool, paid = state.pool, 0
         for b in range(cfg.blocks_per_epoch):
@@ -263,7 +266,7 @@ def run(cfg: Config, scfg: StrategyConfig) -> tuple[Population, list[EpochRecord
             providers=n_prov, service_per_provider_lgo=per_prov,
             claims_paid=paid, reward_per_claim=reward_per_claim,
             pool_lgo=cfg.to_lgo(state.pool),
-            pow_reward_per_block=claims_found * reward_per_claim,
+            pow_reward_per_block=claims_found * np.int64(reward_per_claim),
         ))
         est = est.update(blocks, cfg)
